requirements.txt
aiodns==2.0.0
aiohttp==3.6.1
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptg==0.2.post0
idna==2.8
maxminddb==1.4.1
multidict==4.5.2
proxybroker==0.3.2
pyaes==1.6.1
pyasn1==0.4.6
pycares==3.0.0
pycparser==2.19
PySocks==1.7.0
requests==2.22.0
rsa==4.0
Telethon==1.9.0
urllib3==1.25.3
yarl==1.3.0
